Pull link changer modal into editor-agnostic component
